Adds render to index.js
